# [hail][performance] Deduplicate inlined IRs in `annotate(**thing)`